+++ src/main/java/com/natlowis/games/game/connectfour/BoardConnectFour.java
@@ -1,8 +1,8 @@
 package com.natlowis.games.game.connectfour;
 
 import com.natlowis.games.game.Type;
-import com.natlowis.games.game.interfaces.Board;
-import com.natlowis.games.game.interfaces.Piece;
+import com.natlowis.games.game.interfaces.games.Board;
+import com.natlowis.games.game.interfaces.games.Piece;
 import com.natlowis.games.game.naughtsandcrosses.PieceNaughtsAndCrosses;
 
 /**
