+++ src/main/java/com/natlowis/games/game/naughtsandcrosses/BoardNaughtsAndCrosses.java
@@ -1,8 +1,8 @@
 package com.natlowis.games.game.naughtsandcrosses;
 
 import com.natlowis.games.game.Type;
-import com.natlowis.games.game.interfaces.Board;
-import com.natlowis.games.game.interfaces.Piece;
+import com.natlowis.games.game.interfaces.games.Board;
+import com.natlowis.games.game.interfaces.games.Piece;
 
 /**
  * This will hold the game used in Naughts and Crosses
