+++ src/main/java/com/natlowis/games/game/naughtsandcrosses/PieceNaughtsAndCrosses.java
@@ -1,7 +1,7 @@
 package com.natlowis.games.game.naughtsandcrosses;
 
 import com.natlowis.games.game.Type;
-import com.natlowis.games.game.interfaces.Piece;
+import com.natlowis.games.game.interfaces.games.Piece;
 
 /**
  * This creates the {@link Piece} for Naughts and Crosses
