+++ src/main/java/com/natlowis/games/ui/cli/Run.java
@@ -1,7 +1,7 @@
 package com.natlowis.games.ui.cli;
 
 import com.natlowis.games.game.connectfour.PlayConnectFourAi;
-import com.natlowis.games.game.interfaces.Play;
+import com.natlowis.games.game.interfaces.games.Play;
 import com.natlowis.games.game.naughtsandcrosses.PlayNaughtsAndCrosses;
 import com.natlowis.games.game.naughtsandcrosses.PlayNaughtsAndCrossesAi;
